Handle various conditions and cleanup.
Added code to handle deletion of the default boot image.  This now
succeeds so long as the default boot image is not the running image.
The running image still can not be deleted.

Correctly handle case where image being deleted resides before the
default boot image in the grub config file.  Previously this would
either switch the default boot image to the one after it, or leave the
default pointing to a non-existant entry.

Added the Vyatta license.

Re-organized and added comments to make the code easier to maintain.