Bugfix 5238: Prevent the running image from being deleted
In order to prevent the running image from being deleted, we now get
the image name of the running system from the kernel boot path given
by /proc/cmdline.  Since the menu of names of images to delete comes
from the grub config file, this is guranteed to exactly match the grub
entry that was used to boot.

Also fix the help text for "delete system image" command.