Add console keymap reconfigure op-mode command
Bug 7448
This adds template for the Debian way to reconfigure console keymap.

Note: although this is an op-mode command, it really persists across
reboot since it changes the package keymap. There isn't a good method
to put this in Vyatta config since the debian package configuration
is interactive and prompts user for keyboard type and options.