op-mode: T3582: Fix the function of "delete log file" command
Added the compeletion help to show the list of the files and also changed
the command to delete the file as the existing command was deleting the file
and not recreating the file to write the new logs. So modified the command
so that it will delete the content of the file.