T5411: remove old background monitor implementation
Old Vyatta had a possibility to issue log entries to the users console based
on a service.

Examples:

* monitor dhcp background start | stop
* monitor lldp background start | stop
* monitor openvpn background start | stop

Somehow they not seem to work properly anymore (we are not certain if they
ever did work properly)

The maintainers decided to drop the op-mode CLI nodes from vyatta-op package in
order to get rid of all the legacy code.

In a subsequent task we might wan't to re-add this capability under the more
"industry standard" command terminal monitor.